@@ -1,6 +1,6 @@
-export { Api, ApiError, createApi } from './src/api'
-export { Observer } from './src/observer'
-export { ZetaPushConnection, ZetaPushClient } from './src/core'
-export { ZetaPushClientConfig } from './src/di'
-export { ZetaPushModule } from './src/zetapush.module'
-export const VERSION = '0.2.1'
+export { Api, ApiError, createApi } from './src/api';
+export { Observer } from './src/observer';
+export { ZetaPushConnection, ZetaPushClient } from './src/core';
+export { ZetaPushClientConfig } from './src/di';
+export { ZetaPushModule } from './src/zetapush.module';
+export const VERSION = '0.4.0';
